refactor: organize each query by folder
