+++ BreastCancer.Tests/Unit/UpdatePostCommandHandlerTests.cs
@@ -2,7 +2,7 @@ using AutoMapper;
 using BreastCancer.Community.DTO.request;
 using BreastCancer.Community.DTO.response;
 using BreastCancer.Community.Exceptions;
-using BreastCancer.Community.Features.Posts;
+using BreastCancer.Community.Features.UpdatePost;
 using BreastCancer.Community.Services.Interface;
 using BreastCancer.Context;
 using BreastCancer.Enum;
