+++ Community/Features/DeletePost/DeletePostCommand.cs
@@ -1,5 +1,5 @@
 using MediatR;
 
-namespace BreastCancer.Community.Features.Posts;
+namespace BreastCancer.Community.Features.DeletePost;
 
 public sealed record DeletePostCommand(int PostId, string RequesterId, IReadOnlyCollection<string> Roles) : IRequest<Unit>;
