+++ Community/Features/DeletePost/DeletePostCommandHandler.cs
@@ -4,7 +4,7 @@ using BreastCancer.Context;
 using Microsoft.EntityFrameworkCore;
 using MediatR;
 
-namespace BreastCancer.Community.Features.Posts;
+namespace BreastCancer.Community.Features.DeletePost;
 
 public sealed class DeletePostCommandHandler : IRequestHandler<DeletePostCommand, Unit>
 {
