+++ Community/Features/GetPost/GetPostQuery.cs
@@ -1,6 +1,6 @@
 using BreastCancer.Community.DTO.response;
 using MediatR;
 
-namespace BreastCancer.Community.Features.Posts;
+namespace BreastCancer.Community.Features.GetPost;
 
 public sealed record GetPostQuery(int PostId, string RequesterId, IReadOnlyCollection<string> Roles) : IRequest<PostDTO>;
