+++ Community/Features/UpdatePost/UpdatePostCommand.cs
@@ -2,6 +2,6 @@ using BreastCancer.Community.DTO.request;
 using BreastCancer.Community.DTO.response;
 using MediatR;
 
-namespace BreastCancer.Community.Features.Posts;
+namespace BreastCancer.Community.Features.UpdatePost;
 
 public sealed record UpdatePostCommand(int PostId, UpdatePostDTO Post, string RequesterId) : IRequest<PostDTO>;
