+++ Community/Features/UpdatePost/UpdatePostCommandHandler.cs
@@ -6,7 +6,7 @@ using BreastCancer.Context;
 using Microsoft.EntityFrameworkCore;
 using MediatR;
 
-namespace BreastCancer.Community.Features.Posts;
+namespace BreastCancer.Community.Features.UpdatePost;
 
 public sealed class UpdatePostCommandHandler : IRequestHandler<UpdatePostCommand, PostDTO>
 {
